refactor(hub): extract FleetHost from DetailWindow; FleetDetailPane.initialTab
